Fix the base case for 'Backstage Passes'
As the code was written, it would only work with backstage passes to a
specific concert, 'Backstage passes to a TAFKAL80ETC concert'.

With this change, it does a case-insensitive match against the first 16
characters of the item name to 'backstage passes'. This allows The
Gilded Rose to use the same rules when selling backstage passes for any
concert without a code change.

This is inline with the specificaton document (README.md) as it defines
a general requirement for 'Backstage Passes', not a specific
requirement for 'Backstage passes to a TAFKAL80ETC concert'.